Refactor Manager and Utils
